Make requirements from Pipfile.lock, instead of Pipfile.
If _Pipfile_ is altered manually after running the last
series of tests, the generated _requirements.txt_ is not in
line with the latest run of the environment against which
was tested. The module `pipenv-to-requirements` solves
exactly that problem.
Remark that this seems to be the thought of people over at
the Python Packaging Authority, see the last paragraph of
the issue reported in
https://github.com/pypa/pipfile/issues/7#issue-190667376

> Tools work with Pipfile.lock exclusively. A build system should never
automatically execute a Pipfile to generate a missing or outdated
Pipfile.lock